+++ agent-control/src/sub_agent/on_host/command.rs
@@ -1,7 +1,6 @@
 pub mod command_os;
 pub mod error;
 pub mod executable_data;
-pub mod filesystem_entries;
 pub mod logging;
 pub mod restart_policy;
 pub mod shutdown;
